+++ docs/conf.py
@@ -14,7 +14,7 @@ if True:
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
 project = "dsptoolbox"
-copyright = "2023, Nicolas Franco-Gomez"
+copyright = "2025, Nicolas Franco-Gomez"
 author = "Nicolas Franco-Gomez"
 
 release = __version__